src/requirements.txt
appdirs==1.4.4
asgiref==3.3.4
distlib==0.3.2
Django==3.2.4
django-crispy-forms==1.12.0
django-filter==2.4.0
et-xmlfile==1.1.0
filelock==3.0.12
jdcal==1.4.1
openpyxl==2.6.0
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
virtualenv==20.4.7
xlwt==1.3.0
